+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added new field `national_interest`
+
+### Fixed
+
+- Fixed bug where value for debris lithology was not displayed.
+
 ## v2.0.441 - 2023-06-12
 
 ### Added